Import of environment variables fixed
Due to the nature of the CircleCI API a straight import is not possible
as the "ID" of the resource is only the name. For the rest of the import
to work you need the organisation and the project name.

This encodes them into the ID so an import can happen successfully.