Improved import functionality (#27)
Due to the nature of the CircleCI API a straight import is not possible
as the "ID" of the resource is only the name. For the rest of the import
to work you need the organization and the project name.

This encodes them into the ID so an import can happen successfully.

This allows for the project name to include '.'s. As environment vars
can't have them according to the CircleCI spec and orgs are not allow
'.'s in GitHub, this should be relatively safe